+++ keyboards/planck/keymaps/prog_qgmlwb/src/base.h
@@ -7,13 +7,12 @@
 
 #include "config.h"
 
-#include "base/keymap_norwegian.h"
-#include "base/functions.h"
-#include "base/layout.h"
+#include "keymap/norwegian.h"
+#include "keymap/layouts.h"
+
+#include "util/functions.h"
 
 #include "hardware/left_space.h"
 #include "hardware/swap_gui_alt.h"
 
 #include "macros/macros.h"
-
-extern const uint16_t PROGMEM keymaps [][MATRIX_ROWS][MATRIX_COLS];
