+++ keyboards/planck/keymaps/prog_qgmlwb/src/keymap/norwegian.h
