+++ keyboards/planck/keymaps/prog_qgmlwb/src/util/functions.c
